pool: adjust layout
This PR splits pool into several submodels. `Runner` is also moved into
pool as it's an implement details of the pool, should not be exposed to
users in most cases.
